requirements.txt
fastapi==0.104.1
twilio==8.11.0
openai==1.3.9
python-dotenv==1.0.0
hathor-wallet-lib==0.56.0
hathor-api-client==0.1.1
apscheduler==3.10.4
uvicorn==0.24.0
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
